Finished the Material Controller
Material Controller Is now complete, with the ability to create a MI (Material Instance). This will give you the ability 2 switch out textures and and turn on cell bombing to edit textures to remove repetitiveness of tilling textures. it also have a texture switch out to switch to a larger texture file further away from the camera. there is also post process controls as well and color tool to change texture to make it more variations in textures.